Refactored functions to utils
