+++ app/react/Markdown/components/index.js
@@ -5,7 +5,7 @@ import { Icon } from 'UI';
 import Counter from './Counter';
 import ContactForm from './ContactForm';
 import { EntityData } from './EntityData';
-import { Section } from './Section';
+import { EntitySection } from './EntitySection';
 import EntityLink from './EntityLink';
 import ItemList from './ItemList';
 import Repeat from './Repeat';
@@ -67,5 +67,5 @@ export {
   Map,
   Link,
   EntityInfo,
-  Section,
+  EntitySection,
 };
